add uuid in zones
