+++ README.md
@@ -11,8 +11,10 @@ REACT_APP_CLIENT_ID=123 REACT_APP_CLIENT_SECRET=abc456 npm run start
 
 - [x] Disable button while making a PUT call to Strava
 - [x] Add top bar with account currently logged in
-- [ ] Allow to specify multiple zones and save them accross refreshes (localStorage)
-- [ ] Allow lat/lng search through https://nominatim.org/
+- [x] Allow to specify multiple zones and save them accross refreshes (localStorage)
+- [x] Allow lat/lng search through https://nominatim.org/
+- [ ] Do not save uuid in local storage.
+- [ ] Recenter the map when searching for a place.
 - [ ] Allow to specify days and save them accross refreshes (localStorage)
 - [ ] Add backend
 - [ ] Deploy on Heroku
