+++ src/CriteriaModal.js
@@ -2,7 +2,8 @@ import ReactDom from "react-dom";
 import { useEffect, useRef, useState } from "react";
 import "./CriteriaModal.css";
 import { Fragment } from "react";
-import Zone from "./Zone";
+import ZoneSelector from "./ZoneSelector";
+import { v4 as uuidv4 } from "uuid";
 
 let modalWrapper = document.querySelector(".modal-wrapper");
 if (!modalWrapper) {
@@ -21,12 +22,21 @@ const daysOfWeeks = {
   7: "Sunday",
 };
 
+export const defaultZone = {
+  lat: 52.5175672,
+  lng: 13.3981842,
+  radius: 250,
+};
+
 const CriteriaModal = ({ onClose, zones, setZones }) => {
   const backdrop = useRef();
-  const addAZone = () => {
-    setZones((oldZones) => [...oldZones, oldZones[0]]);
+  const addZone = () => {
+    setZones((oldZones) => [
+      ...oldZones,
+      { ...(oldZones[0] || defaultZone), uuid: uuidv4() },
+    ]);
   };
-  const removeAZone = (index) => {
+  const removeZone = (index) => {
     setZones((zones) => zones.filter((_zone, i) => i !== index));
   };
 
@@ -84,96 +94,22 @@ const CriteriaModal = ({ onClose, zones, setZones }) => {
               </Fragment>
             );
           })}
-          {zones.map((zone, index) => {
-            const { lat, lng, radius } = zone;
-            const updateZone = (zone) => {
-              setZones((zones) =>
-                zones.map((oldZone, i) =>
-                  i === index ? { ...oldZone, ...zone } : oldZone
-                )
-              );
-            };
-            const handleSubmitLocation = async (event) => {
-              event.preventDefault();
-              const searchLocation = event.target.searchLocation.value;
-              const response = await fetch(
-                `https://nominatim.openstreetmap.org/search/${searchLocation}?format=json&addressdetails=1&limit=1&polygon_svg=1`
-              );
-              const data = await response.json();
-              debugger;
-              updateZone({
-                lng: parseFloat(data[0].lon, 10),
-                lat: parseFloat(data[0].lat, 10),
-              });
-            };
-
-            return (
-              <div className="criteria-modal-zone">
-                <div>
-                  <h2 className="text-center">Create a zone - zone {index}</h2>
-                  <p>
-                    Type here the latitude and longitude, or click on the map
-                  </p>
-                  <label className="display-block">
-                    Latitude{" "}
-                    <input
-                      step="0.1"
-                      min={0}
-                      type="number"
-                      value={lat}
-                      onChange={(e) =>
-                        updateZone({ lat: parseFloat(e.target.value, 10) })
-                      }
-                    />
-                  </label>
-                  <label className="display-block">
-                    Longitude{" "}
-                    <input
-                      step="0.1"
-                      min={0}
-                      type="number"
-                      value={lng}
-                      onChange={(e) =>
-                        updateZone({ lng: parseFloat(e.target.value, 10) })
-                      }
-                    />
-                  </label>
-                  <label className="display-block">
-                    Location{" "}
-                    <form onSubmit={(e) => handleSubmitLocation(e)}>
-                      <input
-                        type="text"
-                        placeholder="Search..."
-                        name="searchLocation"
-                      ></input>
-                      <button type="submit">&#x1F50D;</button>
-                    </form>
-                  </label>
-                  <label className="display-block">
-                    Radius {radius} meters{" "}
-                    <input
-                      step="1"
-                      min={0}
-                      max={1000}
-                      type="range"
-                      value={radius}
-                      onChange={(e) =>
-                        updateZone({ radius: parseFloat(e.target.value, 10) })
-                      }
-                    />
-                  </label>
-
-                  {index > 0 && (
-                    <button onClick={() => removeAZone(index)}>
-                      Remove zone
-                    </button>
-                  )}
-                </div>
-                <Zone zone={zone} index={index} updateZone={updateZone} />
-              </div>
-            );
-          })}
-          <button onClick={() => addAZone()}>Add a zone</button>
+          {zones.map((zone, index) => (
+            <ZoneSelector
+              key={zone.uuid}
+              zone={zone}
+              index={index}
+              removeZone={() => removeZone(index)}
+              updateZone={(zone) => {
+                setZones((zones) =>
+                  zones.map((oldZone, i) =>
+                    i === index ? { ...oldZone, ...zone } : oldZone
+                  )
+                );
+              }}
+            />
+          ))}
+          <button onClick={addZone}>Add a zone</button>
         </div>
       </div>
     </div>
